tplink-hs: Move tplink_hs_ops to protocol.h
